refactor: better types without as assertion
