+++ .github/workflows/codeql.yml
@@ -28,7 +28,7 @@ jobs:
           - javascript
 
     steps:
-      - name: Checkout repository
+      - name: Checkout Repo
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
 
       - name: Initialize CodeQL
