--- src/helper.ts
@@ -1,4 +1,0 @@
-const _slice = (Array.prototype as unknown[]).slice // type-coverage:ignore-line
-
-export const slice = <T>(items: { [index: number]: T; length: number }) =>
-  _slice.call(items) as T[]
